fewer casts

+++ src/history.rs
@@ -31,8 +31,8 @@ impl HashHist {
         //hist[self.pointer -1] = pos.hash
         //hist[self.pointer -3] = pos_2movesago.hash != pos.hash
         //=> Start at 5, stride 2
-        for i in (5..=(pos.mr50 + 1).min(self.pointer as u8)).step_by(2) {
-            if self.hist[self.pointer - i as usize] == pos.hash {
+        for i in (5..=(pos.mr50 as usize + 1).min(self.pointer)).step_by(2) {
+            if self.hist[self.pointer - i] == pos.hash {
                 return true;
             }
         }
